\name{fdaRNA-package}
\alias{fdaRNA-package}
\alias{fdaRNA}
\docType{package}
\title{RNA-seq and microarray preprocessing
\packageTitle{fdaRNA}
}
\description{
This package is dedicated to two central aspects of gene expression preprocessing: normalization and outlier detection. In this package, those methodologies are based on statistical depth. The outlier detection methodology can be generally applied.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fdaRNA}
\packageIndices{fdaRNA}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fdaRNA}

Maintainer: \packageMaintainer{fdaRNA}
}
\references{
Nieto-Reyes A, Cabrera J. Statistical depth based normalization and outlier detection of gene expression data. Preprint.
}
\keyword{nonparametric}
\keyword{multivariate}
\keyword{robust}

