% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.boxplot.R
\name{normalization.boxplot}
\alias{normalization.boxplot}
\title{Boxplot of gene expressions}
\usage{
normalization.boxplot(
  ar,
  namesN = T,
  before = F,
  after = T,
  par.r = F,
  par.c = F,
  arg1 = "Before",
  arg2 = "After"
)
}
\arguments{
\item{ar}{Cell files from Affy micrarray experiment: a matrix where the columns are the sample 
elements, RNA-seq or microarray, and the rows the variables (the genes).}

\item{namesN}{If TRUE, the labels of each RNA-seq's or microarray are substituted by a number 
corresponding to its position in data matrix. The default is TRUE.}

\item{before}{If TRUE, the "log(ar+1)" transform of the data before the normalization step is
plotted. The default is FALSE.}

\item{after}{If TRUE, the "log(ar+1)" transform of the data after the normalization step is plotted. 
The default is TRUE.}

\item{par.r}{If TRUE, and both "before" and "after" are also TRUE, the "before" and "after" boxplots 
are displayed in a row. The default is FALSE.}

\item{par.c}{IIf TRUE, and both "before" and "after" are also TRUE, the "before" and "after" boxplots 
are displayed in a column. The default is FALSE.}

\item{arg1}{Overall title for the boxplot before the normalization step. The default is \emph{Before}.}

\item{arg2}{Overall title for the boxplot after the normalization step. The default is \emph{After}.}
}
\value{
The boxplot of the "log( +1)" transform of the data before and/or after the normalization step.
}
\description{
Boxplot of the "log(ar+1)" transform of the data before and after the statistical functional 
depth based normalization. This is commonly carried out to show the adequacy of the procedure.
}
\details{
Boxplot of the "log(ar+1)" transform of the data before and after the normalization step. This is 
commonly carried out to show the adequacy of the procedure.
}
\examples{
## Computes the normalization of the RNA-seq's in the Airway dataset.
N = normalization(Airway)

## Displays in a row the boxplots of the "log(+1)" transform 
## of the data before and after normalization.

normalization.boxplot(Airway, before = TRUE, par.r = TRUE)

}
\references{
Nieto-Reyes A, Cabrera J. Statistical depth based normalization and outlier detection of gene expression 
data. Preprint.
}
\author{
A. Nieto and J. Cabrera
}
\keyword{multivariate}
\keyword{non-parametric.}
\keyword{robust}
