% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{Statistical depth based normalization}
\usage{
normalization(x)
}
\arguments{
\item{x}{Cell files from Affy micrarray experiment: a matrix where the columns are the sample 
elements, RNA-seq or microarray, and the rows the variables (the genes).}
}
\value{
It returns a matrix of the same dimension as the argument, which consists of the
normalized data.
}
\description{
Gene expression normalization equates the scales at which the various gene expressions 
have been measured. This function follows the procedure in Nieto-Reyes and Cabrera (2020) 
that equates to the statistical sample functional depth in Nieto-Reyes (2011).
}
\examples{
## Computes the normalization of the RNA-seq's in the Airway dataset.
N = normalization(Airway)

## Displays in a row the boxplots of the "log( +1)" transform of 
## the data before and after normalization.
normalization.boxplot(Airway, before = TRUE, par.r = TRUE)

}
\references{
Amaratunga D, Cabrera J, Shkedy Z. Exploration and analysis of DNA microarray and other high 
dimensional data. J. Wiley & Sons, 2014. 
Nieto-Reyes A. (2011) On the Properties of Functional Depth. In: Ferraty F. (eds) 
Recent Advances in Functional Data Analysis and Related Topics. Contributions to Statistics. 
Physica-Verlag HD.
Nieto-Reyes A, Cabrera J. Statistical depth based normalization and outlier 
detection of gene expression data. Preprint.
}
\author{
A. Nieto and J. Cabrera
}
\keyword{multivariate}
\keyword{robust}
