% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.order.R
\name{plot_order}
\alias{plot_order}
\title{Plots ordered datasets}
\usage{
plot_order(
  ar,
  x = 1:nrow(ar),
  tp = "l",
  xl = range(x),
  yl = c(min(ar[x, ]), max(ar[x, ])),
  xlb = "",
  ylb = "",
  m = "",
  ...
)
}
\arguments{
\item{ar}{Cell files from Affy micrarray experiment: a matrix where the columns are the sample 
elements, RNA-seq or microarray, and the rows the variables (the genes).}

\item{x}{A vector that is subset of the natural numbers up to the number of variables. 
It represents the domain, of the X-axis, to use for the zoomed plot. The default is the whole domain.}

\item{tp}{The type of plot that should be drawn. The default is "l" for lines.}

\item{xl}{A vector with the limits of the X-axis. The default is the range of "x".}

\item{yl}{A vector with the limits of the Y-axis. The default is the range of "ar" when restricted 
to the domain "x".}

\item{xlb}{X-axis label for the  multidimensional scaling and/or spectral map plots. The default 
is empty. It is only used if "MS" or/and "SM" is/are set to TRUE.}

\item{ylb}{Y-axis label for the  multidimensional scaling and/or spectral map plots. The default 
is empty. It is only used if "MS" or/and "SM" is/are set to TRUE.}

\item{m}{Main title plot. The default is empty}

\item{...}{Additional arguments to be passed to the multidimensional scaling and/or spectral map 
plots, such as graphical parameters. It is only used if "MS" or/and "SM" is/are set to TRUE.}
}
\value{
does not return
}
\description{
In the process of normalization, statistical depth based or quantile, each element of the sample 
is ordered in ascending order. This function plots the sample elements in this process.
}
\details{
This function plots allows to plot a zoom of the domain. The color shows the deepness of the 
sample elemnt with respect to the sample, using \emph{prof.funct}. The colours go from blue 
meaning low depth to red meaning high depth, through green and yellow.
}
\examples{
# Plots the Sialin data set after 6 hours of gestation, in the ascending order used 
# in the process of normalization
plot_order(Sialin6)

# A zoom of previous plot in the interval [20000 , 30000]
plot_order(Sialin6, c(20000:30000))

}
\references{
Nieto-Reyes A, Cabrera J. Statistical depth based normalization and outlier detection of gene 
expression data. Preprint.
}
\seealso{
\code{\link{normalization}} and \code{\link{prof.funct}}
}
\author{
Alicia Nieto-Reyes and Javier Cabrera
}
\keyword{multivariate}
\keyword{non-parametric.}
\keyword{robust}
