% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaRNA-package.R
\docType{package}
\name{fdaRNA-package}
\alias{fdaRNA-package}
\alias{fdaRNA}
\title{The 'fdaRNA' package.}
\description{
This package is dedicated to two central aspects of gene expression preprocessing: 
normalization and outlier detection. In this package, those methodologies are based 
on statistical depth. The outlier detection methodology can be generally applied.
}
\references{
Nieto-Reyes A, Cabrera J. Statistical depth based normalization and outlier detection of 
gene expression data. Preprint
}
\keyword{multivariate,}
\keyword{non-parametric,}
\keyword{robust}
