% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdrma.R
\name{fdrma}
\alias{fdrma}
\title{Functional Data Robust Microarray Analysis}
\usage{
fdrma(
  slc0cel,
  method = c("fdn", "myrma"),
  mod = c("MP", "MM", "LS"),
  no.cores = floor(detectCores())
)
}
\arguments{
\item{slc0cel}{Cell files from Affy micrarray experiment.}

\item{method}{Normalization method: either "fdn" (funcional data normalization) or 
"qn" quantile normalization.}

\item{mod}{one of the following: "MP"= median polish, 
"MM"= Tukey's biweight M-estimator,"M"= Huber M-estimator,"LS"= Least squares.}

\item{no.cores}{default is floor(detectCores())-1}
}
\value{
This function returns an array of gene expressions with G genes by n samples, 
that summarizes the probes of the original datasets.
}
\description{
This function performs robust microarray analysis for
a sample of microarray cell files from a microarray
experiment with Affymetric chips.
}
\details{
functional data  normalization instead of quantile  normalization. In addition the user may 
specify a robust regression M-estimator of regression ("M" or "MM") instead of one step median 
polish that was used in the original RMA. MM-estimator is an M-estimator that uses the psi function 
given by Tukey's biweight function. The M-estimator uses Huber's psi function.
}
\examples{
\dontrun{
library(affy)
# Path to file with cel files from experiment
fns <- list.celfiles(path="cel_filesday0",full.names=TRUE)
# Read Affy files into R
slc0cel <- ReadAffy(filenames=fns)

# Get rma file and fdn files with gene expressions
slc0.fdn= fdrma(slc0cel, method = "fdn", mod = "MM")
slc0.rma= fdrma(slc0cel, method = "rma")
# Compare the functional forms of the gene expressions
plot(sort(slc0.rma[,1]),sort(slc0.fdn[,1]),type="l",col=2)
}

}
\author{
A. Nieto and J. Cabrera
}
\keyword{multivariate}
\keyword{robust}
